Conflicts:
	.gitignore
	src/Three20/Three20.xcodeproj/project.pbxproj
	src/Three20Core/Three20Core.xcodeproj/project.pbxproj
	src/Three20Network/Three20Network.xcodeproj/project.pbxproj
	src/Three20Style/Three20Style.xcodeproj/project.pbxproj
	src/Three20UI/Sources/TTTableViewController.m
	src/Three20UI/Sources/TTThumbsDataSource.m
	src/Three20UI/Three20UI.xcodeproj/project.pbxproj
	src/Three20UICommon/Three20UICommon.xcodeproj/project.pbxproj
	src/Three20UINavigator/Three20UINavigator.xcodeproj/project.pbxproj

+++ src/Three20UICommon/Sources/UIViewControllerAdditions.m
@@ -1,5 +1,5 @@
 //
-// Copyright 2009-2010 Facebook
+// Copyright 2009-2011 Facebook
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -346,7 +346,7 @@ static const NSTimeInterval kGarbageCollectionInterval = 20;
       NSInteger retainCount = [controller retainCount] - 1;
 
       TTDCONDITIONLOG(TTDFLAG_CONTROLLERGARBAGECOLLECTION,
-                      @"Retain count for %X is %d", controller, retainCount);
+                      @"Retain count for %X is %d", (unsigned int)controller, retainCount);
 
       if (retainCount == 1) {
         // If this fails, you've somehow added a controller that doesn't use
@@ -370,7 +370,7 @@ static const NSTimeInterval kGarbageCollectionInterval = 20;
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 - (void)unsetCommonProperties {
   TTDCONDITIONLOG(TTDFLAG_CONTROLLERGARBAGECOLLECTION,
-                  @"Unsetting this controller's properties: %X", self);
+                  @"Unsetting this controller's properties: %X", (unsigned int)self);
 
   self.superController = nil;
   self.popupViewController = nil;
